refactor

+++ ts/Casino.ts
@@ -73,7 +73,7 @@ class Casino {
         let ageString: string = this.userInputHTMLElement.value;
         if (ageString != null) {
             if (this.validateAge(ageString)) {
-                this.setPlayerAge(parseFloat(ageString));
+                this.setPlayerAge(parseInt(ageString));
                 this.appendToDisplay("How much money do you have?", true);
                 this.setOnClickAttributeOfSubmitButton("getUserMoney()");
             }
